requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
banks==2.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==3.1
dataclasses-json==0.6.7
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
faiss-cpu==1.13.2
filelock==3.20.2
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
griffe==1.15.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.2.4
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-openai==1.1.6
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.1
llama-cloud==0.1.45
llama-cloud-services==0.6.88
llama-index-core==0.14.12
llama-index-instrumentation==0.4.2
llama-index-workflows==2.11.7
llama-parse==0.6.88
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.0
onnxruntime==1.23.2
openai==2.14.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pillow==12.1.0
platformdirs==4.5.1
propcache==0.4.1
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyMuPDF==1.26.7
pyreadline3==3.5.4
python-dotenv==1.2.1
PyYAML==6.0.3
rank-bm25==0.2.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0